pyproject.toml
[project]
name = "ai-assistant-backend"
version = "0.1.0"
requires-python = ">=3.12"

[dependency-groups]
dependencies = [
    
]

dev = [
    "mypy==1.16.0",
    "ruff==0.11.13",
    "langdetect==1.0.9",
    "colorama==0.4.6",
    "langchain==0.3.27",
    "langchain-community==0.3.27",
    "langchain-mongodb==0.6.2",
    "motor==3.7.1",
    "pathlib2==2.3.7.post1",
    "fastapi==0.116.1",
    "uvicorn==0.35.0",
    "tavily-python==0.7.10",
    "langchain-ollama==0.3.6",
    "llama-cpp-python",
    {include-group = "ingestion"},
]

ingestion = [
    "PyMuPDF==1.26.3",
    "docx==0.2.4",
    "pymongo==4.14.0",
    "google-cloud==0.34.0",
    "google-cloud-storage==3.2.0",
    "sentence-transformers==5.1.0",
    "langchain-huggingface==0.3.1",
    "google-cloud-secret-manager==2.24.0",
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
# B for enable flake8-bugbear
select = ["E4", "E7", "E9", "F", "B"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"


[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
